moving the resource classes and user-related classes to a single package(models), moving the serialization and deserialization classes and interfaces to a single package(serde)
