+++ database/serde/Deserializer.java
@@ -1,4 +1,4 @@
-package database.deserializer;
+package database.serde;
 
 public class Deserializer implements Deserializable {
 }
