+++ models/ProfessionType.java
@@ -1,4 +1,4 @@
-package user.professor;
+package models;
 
 public enum ProfessionType {
     CALCULUS, DISCRETE_MATH, LINEAR_ALGEBRA, HISTORY, LITERATURE, LANGUAGE
