+++ models/StudentType.java
@@ -1,4 +1,4 @@
-package user.student;
+package models;
 
 public enum StudentType {
     FRESHMAN, SOPHOMORE, JUNIOR, SENIOR
